+++ entries_holder.go
@@ -60,58 +60,51 @@ type (
 
 	Term struct {
 		FieldID uint64
-		IDValue uint64
+		Value   string
 	}
+
 	// DefaultEntriesHolder EntriesHolder implement base on hash map holder map<key, Entries>
 	// 默认容器,目前支持表达式最大256个field; 支持多个field复用默认容器; 见:Key编码逻辑
 	// 如果需要打破这个限制,可以自己实现容器.
 	DefaultEntriesHolder struct {
-		debug     bool
-		maxLen    int64 // max length of Entries
-		avgLen    int64 // avg length of Entries
-		plEntries map[Term]Entries
-
-		Parser      parser.FieldValueParser
-		FieldParser map[BEField]parser.FieldValueParser
+		debug       bool
+		maxLen      int64 // max length of Entries
+		avgLen      int64 // avg length of Entries
+		plEntries   map[Term]Entries
+		fieldParser map[BEField]parser.ValueTokenizer
 	}
 
-	Uint64TxData cache.Uint64ListValues
+	StrTokenData struct {
+		cache.StrListValues
+	}
 )
 
-var BetterToCacheMaxItemsCount = 512
-
-func NewTerm(fid, idValue uint64) Term {
-	return Term{FieldID: fid, IDValue: idValue}
+func NewTerm(fid uint64, value string) Term {
+	return Term{FieldID: fid, Value: value}
 }
 
 func (tm Term) String() string {
-	return fmt.Sprintf("<%d,%d>", tm.FieldID, tm.IDValue)
+	return fmt.Sprintf("<%d,%s>", tm.FieldID, tm.Value)
 }
 
 func NewDefaultEntriesHolder() *DefaultEntriesHolder {
 	return &DefaultEntriesHolder{
-		plEntries:   map[Term]Entries{},
-		Parser:      parser.NewCommonParser(),
-		FieldParser: map[BEField]parser.FieldValueParser{},
+		plEntries:   make(map[Term]Entries),
+		fieldParser: make(map[BEField]parser.ValueTokenizer),
 	}
 }
 
-func (txd *Uint64TxData) BetterToCache() bool {
-	return len(txd.Values) > BetterToCacheMaxItemsCount
-}
-
-func (txd *Uint64TxData) Encode() ([]byte, error) {
-	protoMsg := (*cache.Uint64ListValues)(txd)
-	return proto.Marshal(protoMsg)
+func (std *StrTokenData) Encode() ([]byte, error) {
+	return proto.Marshal(&std.StrListValues)
 }
 
 // DecodeFieldIndexingData decode data; used for building progress cache
 func (h *DefaultEntriesHolder) DecodeFieldIndexingData(data []byte) (IndexingData, error) {
 	if len(data) == 0 {
-		return &Uint64TxData{Values: nil}, nil
+		return &StrTokenData{}, nil
 	}
-	txData := &Uint64TxData{}
-	err := proto.Unmarshal(data, (*cache.Uint64ListValues)(txData))
+	txData := &StrTokenData{}
+	err := proto.Unmarshal(data, &txData.StrListValues)
 	return txData, err
 }
 
@@ -128,8 +121,8 @@ func (h *DefaultEntriesHolder) DumpInfo(buffer *strings.Builder) {
 		"maxEntriesLen": h.maxLen,
 		"avgEntriesLen": h.avgLen,
 	}
-	for field, idGen := range h.FieldParser {
-		summary[fmt.Sprintf("field#%s#parser", field)] = idGen.Name()
+	for field := range h.fieldParser {
+		summary[fmt.Sprintf("field#%s#parser", field)] = "custom"
 	}
 	buffer.WriteString(util.JSONPretty(summary))
 }
@@ -144,11 +137,21 @@ func (h *DefaultEntriesHolder) DumpEntries(buffer *strings.Builder) {
 	}
 }
 
-func (h *DefaultEntriesHolder) GetParser(field BEField) parser.FieldValueParser {
-	if p, ok := h.FieldParser[field]; ok {
+func (h *DefaultEntriesHolder) GetTokenizer(field BEField) parser.ValueTokenizer {
+	if p, ok := h.fieldParser[field]; ok {
 		return p
 	}
-	return h.Parser
+	return parser.NewDefaultTokenizer()
+}
+
+// RegisterFieldTokenizer registers a custom tokenizer for a specific field.
+// If nil is passed, the default tokenizer will be used.
+func (h *DefaultEntriesHolder) RegisterFieldTokenizer(field BEField, tokenizer parser.ValueTokenizer) {
+	if tokenizer == nil {
+		delete(h.fieldParser, field)
+		return
+	}
+	h.fieldParser[field] = tokenizer
 }
 
 func (h *DefaultEntriesHolder) CompileEntries() error {
@@ -157,14 +160,14 @@ func (h *DefaultEntriesHolder) CompileEntries() error {
 }
 
 func (h *DefaultEntriesHolder) GetEntries(field *FieldDesc, assigns Values) (r EntriesCursors, e error) {
-	var ids []uint64
-	if ids, e = h.GetParser(field.Field).ParseAssign(assigns); e != nil {
+	var values []string
+	if values, e = h.GetTokenizer(field.Field).TokenizeAssign(assigns); e != nil {
 		return nil, e
 	}
-	for _, id := range ids {
-		key := NewTerm(field.ID, id)
+	for _, value := range values {
+		key := NewTerm(field.ID, value)
 		if entries, hit := h.plEntries[key]; hit && len(entries) > 0 {
-			cursor := NewEntriesCursor(NewQKey(field.Field, id), entries)
+			cursor := NewEntriesCursor(NewQKey(field.Field, value), entries)
 			r = append(r, cursor)
 		}
 	}
@@ -174,12 +177,12 @@ func (h *DefaultEntriesHolder) GetEntries(field *FieldDesc, assigns Values) (r E
 func (h *DefaultEntriesHolder) BuildFieldIndexingData(field *FieldDesc, bv *BoolValues) (IndexingData, error) {
 	util.PanicIf(bv.Operator != ValueOptEQ, "default container support EQ operator only")
 
-	// NOTE: ids can be replicated if expression contain cross condition
-	ids, e := h.GetParser(field.Field).ParseValue(bv.Value)
+	// NOTE: values can be replicated if expression contain cross condition
+	values, e := h.GetTokenizer(field.Field).TokenizeValue(bv.Value)
 	if e != nil {
 		return nil, fmt.Errorf("field:%s value:%+v parse fail, err:%s", field.Field, bv, e.Error())
 	}
-	return &Uint64TxData{Values: ids}, nil
+	return &StrTokenData{StrListValues: cache.StrListValues{Values: values}}, nil
 }
 
 func (h *DefaultEntriesHolder) CommitFieldIndexingData(tx FieldIndexingData) error {
@@ -188,12 +191,12 @@ func (h *DefaultEntriesHolder) CommitFieldIndexingData(tx FieldIndexingData) err
 	}
 
 	var ok bool
-	data, ok := tx.Data.(*Uint64TxData)
-	util.PanicIf(!ok, "bad TxData need *Uint64TxData, oops")
+	data, ok := tx.Data.(*StrTokenData)
+	util.PanicIf(!ok, "bad TxData need *StringTxData, oops")
 
-	values := util.DistinctInteger(data.Values)
-	for _, id := range values {
-		key := NewTerm(tx.field.ID, id)
+	values := util.DistinctString(data.StrListValues.Values)
+	for _, value := range values {
+		key := NewTerm(tx.field.ID, value)
 		h.plEntries[key] = append(h.plEntries[key], tx.EID)
 	}
 	return nil